fixed crash on loading session list
session saving now works correctly

+++ app/src/main/java/com/example/distancetracker/topbar/TopBar.kt
@@ -191,7 +191,7 @@ class TopBar(
                 sessionDate.text = session.date
 
                 val sessionDurationTV = newRoute.findViewById<TextView>(R.id.sessionDuration)
-                sessionDurationTV.text = Utility.formatSessionTime(session.duration.toInt())
+                sessionDurationTV.text = Utility.formatSessionTime(session.duration)
 
                 val sessionAvgTV = newRoute.findViewById<TextView>(R.id.sessionAvg)
                 sessionAvgTV.text =
@@ -258,7 +258,7 @@ class TopBar(
         sessionTotalDistance.text = String.format("%.4s km", distance)
 
         val sessionDuration = dialog.findViewById<TextView>(R.id.sessionDuration)
-        sessionDuration.text = Utility.formatSessionTime(route.duration.toInt())
+        sessionDuration.text = Utility.formatSessionTime(route.duration)
 
         val sessionAvg = dialog.findViewById<TextView>(R.id.sessionAvg)
         sessionAvg.text = String.format("%.4s km/h", route.averageSpeed)
